+++ src/main/kotlin/edu/rit/gis/sunspec/models/smdx_00806.kt
@@ -15,6 +15,6 @@ class Model_806 {
                  description="",
                  notes="")
             
-   var BatTBD: UShort? = null
+   var BatTBD: Int? = null
 
 }
